This branch added a test but it does not check individual values on each cell yet
