+++ package.json
@@ -1,0 +1,8 @@
+{
+	"devDependencies": {
+		"jest": "^30.2.0",
+		"ts-jest": "^29.4.5",
+		"@types/jest": "^30.0.0",
+		"typescript": "^5.9.3"
+	}
+}
